requirements.txt
flask
pyyaml
fastapi
uvicorn
requests
httpx
pydantic-settings
# TODO: specify version